ai-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-dotenv==1.0.1
google-generativeai==0.4.1
requests==2.31.0
python-multipart==0.0.9
psycopg2-binary==2.9.9
sqlalchemy==2.0.28
alembic==1.13.1
aiohttp==3.9.3
